move fields to their own file
